Explain that downloads are for stable version
The PDF and TXT releases of the document are automatically generated
from README.md. The PDF/TXT releases are considered stable releases that
happen less frequently than updates to README.md.

This commit adds a paragraph to clarify that the download buttons for
PDF and TXT downloads are meant to download the latest stable release of
the document.

The download button does not appear in the PDF/TXT output files, so this
new paragraph must not appear in the output files as well. As a result,
this commit contains some changes in Makefile too to prevent this
paragraph from appearing in the PDF/TXT output files. This is achieved
by ending the lines of this paragraph with an HTML containing the letter
"x", e.g., "<!-- x -->", and updating the "sed" commands in the Makefile
to delete such lines while generating the output files.