Add make target to update date in every build
The new target "set-date" can be used to automatically update the date
in the version-line of the document. The version-line in the document is
an HTML comment that looks like this:

    <!-- Version 0.1.1-draft (2018-02-19) -->

The version information from the version line is automatically picked by
the build scripts in Makefile and included in the PDF and TXT output
files.

The new target "set-date" automates the update of this version line to
replace the date in it with the current date.

The "build" target is now made to depend on the "set-date" target, so
that the version line gets updated automatically on every build.