Use simple git rebase commands
Before this commit, the document uses elaborate `git rebase` commands
such as `git rebase master TOPIC-BRANCH` to explicitly specify the
branch to check out and rebase on `master`.

The elaborate command may be confusing to beginners. For example, in the
usage of `git rebase <branch1> <branch2>`, it is not immediately obvious
which branch is checked out and which branch it is rebased on. One can
eventually see `man git-rebase` and figure that it is `<branch2>` that
is checked out to rebase on `<branch1>`.

The elaborate command is also often unnecessary. We are often already on
the right branch. For example, a contributor spends most of their time
on the topic branch and therefore `git rebase master` is sufficient to
rebase the topic branch on `master`.

Therefore, this commit replaces commands of the form

    git rebase <branch1> <branch2>

with just

    git rebase <branch1>

in favour of simplicity and clarity.

Further to make it explicit to beginners which branch they need to be on
while executing the simpler `git rebase` command, an explicit command of
the form `git checkout <branch2>` has been added before each `git
rebase`. If the user is already on `<branch2>`, then this command has no
effect.