Simplify updating fork's master with git pull
Prior to this change, the commands to update the fork's main development
branch use:

        git fetch upstream
        git merge upstream/master

With this change, this has been simplified to:

        git pull upstream master

As per the fork and pull request workflow, no new changes are committed
or pushed directly to the fork's master branch. The new changes are
always developed on a topic branch of the fork, merged in the upstream
first, and then pulled to the fork.

Therefore I don't see any advantage of performing `git fetch` and
`git merge` in two different steps when it can be done in a simple
manner with `git pull`. As long as there is nothing being committed
directly to the fork's `master`, the `git pull` command would always
succeed by fast-forwarding the master branch to the latest master from
upstream.